Added routing functionality for handling todos
Files and folders structure for routes now separates different routes for different parts of the app. Created 4 route handlers and a model for todos part. Mongoose now creates separate connections for users and todos DBs, so the logic for connecting to MongoDB now moved for each part of the app to corresponding model file (a bit unsafe but ok for now). Thers is also a bunch of small changes: getCurrentUser middleware moved from current-user route handler to index.ts so that it can run automatically for every incoming request, added additional check for requireAuth middleware, to figure out whether or not there is a user with a given id in DB, and a current-user route handler modified sort of the same way - it now sets currentUser not by a value received from possible cookie, but it checks the existence of a user if there is a cookie and sets currentUser based on this check.