Better error handling with global ErrorsContext
ErrorsContext creates an error div in case of any errors, and makes it available to any component. It also creates a function to update errors and this function can be invoked from any component as well. So now errors state is well controlled throughout the whole frontend in only one place, it simplified Signup and Signin components and useRequest hook.